+++ lib/grainstore/worker/mml-builder-child.js
@@ -1,6 +1,6 @@
 var _ = require('underscore');
 var debug = require('debug')('grainstore:mml-builder-child');
-var MMLBuilder = require('../mml_builder');
+var MMLBuilder = require('./mml_builder');
 
 process.on('message', function(input) {
     debug('Child got MML Builder Context');
