+++ lib/grainstore/worker/mml-builder-worker.js
@@ -1,5 +1,5 @@
 var util = require('util');
-var MMLBuilder = require('../mml_builder');
+var MMLBuilder = require('./mml_builder');
 
 var pool = require('generic-pool');
 var fork = require('child_process').fork;
