+++ lib/grainstore/worker/mml_builder.js
@@ -1,6 +1,6 @@
 var _      = require('underscore');
 var millstone = require('millstone');
-var StyleTrans = require('./style_trans');
+var StyleTrans = require('../style_trans');
 var carto = require('carto');
 var debug = require('debug')('grainstore:mml-builder');
 var debugXml = require('debug')('grainstore:mml-builder:xml');
